requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
mutagen
python-dotenv
pylast